requirements.txt
faster-whisper
sounddevice
numpy
keyboard
pyautogui
pystray
Pillow
pyperclip
